Moved plotting backend classes to new files
