+++ sympy/plotting/backends/textbackend/__init__.py
@@ -1,0 +1,3 @@
+from sympy.plotting.backends.textbackend.text import TextBackend
+
+__all__ = ["TextBackend"]
